unobackend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
boto3==1.34.23; python_version >= '3.8'
botocore==1.34.23; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
coreapi==2.3.3
coreschema==0.0.4
django==5.0.1; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-filter==23.5; python_version >= '3.7'
django-mysql==4.12.0; python_version >= '3.8'
django-storages==1.14.2; python_version >= '3.7'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-spectacular==0.27.1; python_version >= '3.6'
gunicorn==21.2.0; python_version >= '3.5'
idna==3.6; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.3; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
markupsafe==2.1.4; python_version >= '3.7'
mysqlclient==2.2.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pillow==10.2.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.32.1; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.17.1; python_version >= '3.8'
s3transfer==0.10.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2023.4; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.7; python_version >= '3.10'
